requirements.txt
Django==3.2.9
djangorestframework==3.12.4
Pillow==8.4.0
mysqlclient==2.0.3
django-mysql==3.8.1